moving the logger
