+++ Cargo.toml
@@ -11,17 +11,9 @@ name = "ussdframework"
 path = "src/lib.rs"
 crate-type = ["lib"]
 
-
-# [workspace]
-# members = [
-#     "src",
-#     "examples",
-# ]
-
 [dependencies]
 chrono = { version = "0.4", features = ["serde"] }
 config = "0.14.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1.0", features = ["full"] }
-# ussdframework = { path = "." }
