+++ src/lib.rs
@@ -1,26 +1,38 @@
 mod menu;
 pub mod prelude;
-mod screens;
 mod types;
 mod ussd_request;
 mod ussd_response;
+mod ussd_screens;
 mod ussd_service;
 mod ussd_session;
+mod log;
 
 extern crate serde;
 
 use prelude::USSDMenu;
-use screens::process_request;
 use ussd_request::USSDRequest;
 use ussd_response::USSDResponse;
+use ussd_screens::process_request;
 use ussd_session::SessionCache;
 
+/// Represents a USSD application.
 pub struct UssdApp {
     functions_path: String,
     pub session_cache: Box<dyn SessionCache>,
 }
 
 impl UssdApp {
+    /// Creates a new instance of `UssdApp`.
+    ///
+    /// # Arguments
+    ///
+    /// * `functions_path` - The path to the functions used by the USSD application.
+    /// * `session_cache` - The session cache implementation used by the USSD application.
+    ///
+    /// # Returns
+    ///
+    /// A new instance of `UssdApp`.
     pub fn new(functions_path: String, session_cache: Box<dyn SessionCache>) -> Self {
         UssdApp {
             functions_path,
@@ -28,6 +40,16 @@ impl UssdApp {
         }
     }
 
+    /// Runs the USSD application with the given request and screens.
+    ///
+    /// # Arguments
+    ///
+    /// * `request` - The USSD request.
+    /// * `screens` - The USSD menu screens.
+    ///
+    /// # Returns
+    ///
+    /// The USSD response.
     pub fn run(&self, request: USSDRequest, screens: USSDMenu) -> ussd_response::USSDResponse {
         process_request(
             &request,
@@ -37,10 +59,13 @@ impl UssdApp {
         )
     }
 
+    /// Displays the menu to the user.
+    ///
+    /// # Arguments
+    ///
+    /// * `ussd_response` - The USSD response containing the menu message.
     pub fn display_menu(&self, ussd_response: &USSDResponse) {
         // Display the menu to the user
-
-        // For example, you can use the following code to display the menu:
         println!("{}", ussd_response.message);
     }
 }
